Rebased from main; deleted GradeManager
