+++ CanvasPlusPlayground/CourseManager.swift
@@ -10,6 +10,7 @@ import SwiftUI
 @Observable
 class CourseManager {
     var courses = [Course]()
+    var enrollments = [Enrollment]()
 
     func getCourses() async {
         guard let (data, _) = await CanvasService.shared.fetch(.getCourses(enrollmentState: "active")) else {
@@ -21,6 +22,19 @@ class CourseManager {
             self.courses = retCourses
         } else {
             print("Failed to decode file data.")
+        }
+    }
+    
+    func getEnrollments() async {
+        guard let (data, _) = await CanvasService.shared.fetch(.getEnrollments) else {
+            print("Failed to fetch enrollments")
+            return
+        }
+        
+        do {
+            enrollments = try JSONDecoder().decode([Enrollment].self, from: data)
+        } catch {
+            print(error)
         }
     }
 }
